Made some static integer casts in the ball update function to silence compiler about cast warnings.

+++ src/ball.cpp
@@ -33,8 +33,8 @@ void Ball::onUpdate()
 {
   // create an array which contains the movement.
   auto movement = std::array<int, 2>();
-  movement[0] = (mDirection[0] * mVelocity);
-  movement[1] = (mDirection[1] * mVelocity);
+  movement[0] = static_cast<int>(mDirection[0] * mVelocity);
+  movement[1] = static_cast<int>(mDirection[1] * mVelocity);
 
   // apply the movement for the position of the ball.
   mRect.x += movement[0];
